Now ::Modern-style call syntax is implemented (with backward compatibility too!), but it is in it's infancy, so there will be a lot of testing before it goes prime-time.  It looks verrrry sexy.
Several code optimizations.  Now File::Util::Interface::Modern has been
introduced and inherits from File::Util::Interface::Classic for base
methods and functionalities.

The t/extended/*t tests are still under heavy development, and have lots
of commented-out code at the moment, but I want to get this commit done
and push forward in implementing a new feature in list_dir if possible--
I want it to provide the option to return a ref-tree of the entire file
path it's called with.